Moves performance typedefs to separate file
